requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py310h9102076_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.4.26=h06a4308_0
catalogue=2.0.7=py310h06a4308_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.15.0=py310hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310h06a4308_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=36.0.0=py310h9ce1e76_0
cymem=2.0.6=py310h295c915_0
cython-blis=0.7.7=py310h9102076_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
en-core-web-sm=3.3.0=pypi_0
entrypoints=0.4=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
filelock=3.6.0=pyhd3eb1b0_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
future=0.18.2=py310h06a4308_1
glib=2.69.1=h4ff587b_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
huggingface_hub=0.2.1=pyhd3eb1b0_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py310h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py310h06a4308_0
ipython=8.3.0=py310h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
it-core-news-sm=3.3.0=pypi_0
jedi=0.18.1=py310h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jsonschema=4.4.0=py310h06a4308_0
jupyter=1.0.0=py310h06a4308_7
jupyter_client=7.2.2=py310h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py310h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
langcodes=3.3.0=pyhd3eb1b0_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libuuid=1.0.3=h7f8727e_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
libxslt=1.1.34=hc22bd24_0
lxml=4.8.0=py310h1f438cf_0
markupsafe=2.0.1=py310h7f8727e_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py310h7f8727e_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
murmurhash=1.0.7=py310h295c915_0
nbclient=0.5.13=py310h06a4308_0
nbconvert=6.4.4=py310h06a4308_0
nbformat=5.3.0=py310h06a4308_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py310h06a4308_0
networkx=2.7.1=pyhd3eb1b0_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
notebook=6.4.11=py310h06a4308_0
numexpr=2.7.3=py310hd732450_1
numpy=1.21.5=py310hfa59a62_2
numpy-base=1.21.5=py310h9585f30_2
openssl=1.1.1o=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py310h295c915_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathy=0.6.1=py310h06a4308_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=21.2.4=py310h06a4308_0
preshed=3.0.6=py310h295c915_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
pt-core-news-sm=3.3.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py310h295c915_6
pyrsistent=0.18.0=py310h7f8727e_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
pytorch=1.10.2=cpu_py310h6894f24_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py310h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.1.2=h7f8727e_1
regex=2022.3.15=py310h7f8727e_0
requests=2.27.1=pyhd3eb1b0_0
sacremoses=0.0.43=pyhd3eb1b0_0
scikit-learn=1.1.0=pypi_0
scipy=1.8.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py310h06a4308_0
shellingham=1.3.1=pyhd3eb1b0_0
sip=4.19.13=py310h295c915_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smart_open=5.2.1=py310h06a4308_0
soupsieve=2.3.1=pyhd3eb1b0_0
spacy=3.3.0=py310h2571103_0
spacy-legacy=3.0.9=py310h06a4308_0
spacy-loggers=1.0.1=pyhd3eb1b0_0
sqlite=3.38.3=hc218d9a_0
srsly=2.4.3=py310h295c915_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py310h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
thinc=8.0.15=py310h2571103_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=py310hb317417_1
tornado=6.1=py310h7f8727e_0
tqdm=4.64.0=py310h06a4308_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.14.1=pyhd3eb1b0_0
typer=0.4.1=py310h06a4308_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py310h06a4308_0
wasabi=0.9.1=py310h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310h06a4308_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310h06a4308_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.12=h7f8727e_2
